# contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "School Management API",
...
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Read the school settings",
        "description": "Everything in one payload for the settings area (spec 6.2.12). Returns only the `identity` group as of TASK-0005a; later cards extend this same envelope additively with sibling groups.",
        "operationId": "GetSettings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/identity": {
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Update the school's identity",
        "description": "School name, short name, address, phone, email, motto, head teacher name (spec 6.2.3). `timezone` and `abbreviation` are not editable here — timezone is fixed, and the abbreviation has its own endpoint and its own optimistic-concurrency pointer. `expectedVersion` must match the identity group's current `versionNumber` (from `GET /settings`) or the save is rejected `409` before anything is written, and BOTH the winning and the losing attempt are recorded on the audit trail (spec 6.2.11).",
        "operationId": "UpdateSchoolIdentity",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "The value of the __Host-XSRF-TOKEN cookie, echoed verbatim (double-submit CSRF, approved contract delta §5). Obtain it from GET /auth/csrf or from a prior response's Set-Cookie.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated key (UUID v4 recommended), 1-255 visible ASCII characters, no whitespace. Optional. A retry with the same key returns the stored response unchanged and sets the `Idempotency-Replay` response header, rather than repeating the request's effect.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSchoolIdentityCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Idempotency-Replay": {
                "description": "Present and set to \"true\" only when this response is a replay of a prior request that used the same Idempotency-Key, rather than a fresh execution.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsIdentityGroupDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "headers": {
              "Idempotency-Replay": {
                "description": "Present and set to \"true\" only when this response is a replay of a prior request that used the same Idempotency-Key, rather than a fresh execution.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "Idempotency-Replay": {
                "description": "Present and set to \"true\" only when this response is a replay of a prior request that used the same Idempotency-Key, rather than a fresh execution.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "Idempotency-Replay": {
                "description": "Present and set to \"true\" only when this response is a replay of a prior request that used the same Idempotency-Key, rather than a fresh execution.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "headers": {
              "Idempotency-Replay": {
                "description": "Present and set to \"true\" only when this response is a replay of a prior request that used the same Idempotency-Key, rather than a fresh execution.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "headers": {
              "Idempotency-Replay": {
                "description": "Present and set to \"true\" only when this response is a replay of a prior request that used the same Idempotency-Key, rather than a fresh execution.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/config-versions": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "List configuration-version history, newest first",
        "description": "Cursor-paginated per spec 9.5 — never offset. `cursor` is the opaque `nextCursor` from a previous page; omit it for the first page. `pageSize` defaults to 25 and is capped at 100.",
        "operationId": "ListConfigVersions",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPageOfConfigVersionSummaryDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/config-versions/{id}": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Read one configuration version in full, including its snapshot",
        "description": "Includes the full `snapshot` — the whole serialised configuration as of this save (spec 6.2.9), not only the group that changed.",
        "operationId": "GetConfigVersion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigVersionDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CursorPageOfConfigVersionSummaryDto": {
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigVersionSummaryDto"
            },
            "description": "The page of items, newest first. Empty (never null) when there is nothing more to return.",
            "example": [
              {
                "id": "0192f0c4-7c3e-7a1b-9f2d-3b8e5a6c1d40",
                "versionNumber": 3,
                "changedGroup": "Identity",
                "actorAdminId": "0192f0c4-0000-7000-8000-000000000099",
                "createdAtUtc": "2026-08-03T09:30:00+00:00"
              }
            ]
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ],
            "description": "`null` when this is the last page.",
            "example": "MQ=="
          }
        },
        "additionalProperties": false,
        "description": "The cursor-pagination response envelope (spec 9.5). string? CursorPage&lt;TItem&gt;.NextCursor is opaque to the\nclient — it must be echoed back verbatim as the next request's cursor, and never parsed or\nconstructed by hand.",
        "example": {
          "items": [
            {
              "id": "0192f0c4-7c3e-7a1b-9f2d-3b8e5a6c1d40",
              "versionNumber": 3,
              "changedGroup": "Identity",
              "actorAdminId": "0192f0c4-0000-7000-8000-000000000099",
              "createdAtUtc": "2026-08-03T09:30:00+00:00"
            }
          ],
          "nextCursor": "MQ=="
        }
      },
...
      "HttpValidationProblemDetails": {
        "required": [
          "traceId"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ],
            "example": "urn:schoolmanagement:error:sample_record.label_taken"
          },
          "title": {
            "type": [
              "null",
              "string"
            ],
            "example": "Conflict with current state"
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32",
            "example": 409
          },
          "detail": {
            "type": [
              "null",
              "string"
            ],
            "example": "A record with that label already exists."
          },
          "instance": {
            "type": [
              "null",
              "string"
            ],
            "example": "/api/v1/reference/records"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "Label": [
                "Label is required."
              ],
              "PageSize": [
                "PageSize must be at most 100."
              ]
            }
          },
          "errorCode": {
            "type": "string",
            "description": "Stable, machine-readable error code. Clients branch on this, never on `detail`. Absent when this response was produced directly by the framework rather than by this API's own result mapping."
          },
          "traceId": {
            "type": "string",
            "description": "Correlation id for this specific response occurrence. Present on every error response; quote it when reporting a problem."
          }
        },
        "additionalProperties": false,
        "description": "An RFC 9457 problem response for a validation failure, returned with status 422. Extends the standard problem shape with `errors`: an object keyed by request property name, whose values are the messages for that property, suitable for attaching to form fields. A 400 (rather than 422) means the request itself could not be parsed.",
        "example": {
          "type": "urn:schoolmanagement:error:request.validation_failed",
          "title": "Validation failed",
          "status": 422,
          "detail": "One or more validation errors occurred.",
          "instance": "/api/v1/reference/records",
          "errorCode": "request.validation_failed",
          "traceId": "0af7651916cd43dd8448eb211c80319c",
          "errors": {
            "Label": [
              "Label is required."
            ],
            "PageSize": [
              "PageSize must be at most 100."
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "required": [
          "traceId"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ],
            "example": "urn:schoolmanagement:error:sample_record.label_taken"
          },
          "title": {
            "type": [
              "null",
              "string"
            ],
            "example": "Conflict with current state"
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32",
            "example": 409
          },
          "detail": {
            "type": [
              "null",
              "string"
            ],
            "example": "A record with that label already exists."
          },
          "instance": {
            "type": [
              "null",
              "string"
            ],
            "example": "/api/v1/reference/records"
          },
          "errorCode": {
            "type": "string",
            "description": "Stable, machine-readable error code. Clients branch on this, never on `detail`. Absent when this response was produced directly by the framework (for example a model-binding failure or an authentication challenge) rather than by this API's own result mapping."
          },
          "traceId": {
            "type": "string",
            "description": "Correlation id for this specific response occurrence. Present on every error response; quote it when reporting a problem."
          },
          "lockedUntil": {
            "type": "string",
            "description": "UTC time the account's lockout ends (spec 6.1.11: five failed attempts locks it for fifteen minutes). Present only on the `423 Locked` response `POST /auth/sign-in` returns when the SUBMITTED password is correct but the account is currently locked — never on any other problem response.",
            "format": "date-time",
            "example": "2026-08-03T09:30:00+00:00"
          }
        },
        "additionalProperties": false,
        "description": "An RFC 9457 problem response. Returned for every error. Branch on the `errorCode` extension member — it is stable — and never on `detail`, which is human-readable prose that may be reworded. `traceId` identifies this specific occurrence in the server logs; quote it when reporting a problem. `type` is a stable URN of the form `urn:schoolmanagement:error:<code>`.",
        "example": {
          "type": "urn:schoolmanagement:error:sample_record.label_taken",
          "title": "Conflict with current state",
          "status": 409,
          "detail": "A record with that label already exists.",
          "instance": "/api/v1/reference/records",
          "errorCode": "sample_record.label_taken",
          "traceId": "0af7651916cd43dd8448eb211c80319c"
        }
      },
...
      "SettingsDto": {
        "required": [
          "identity"
        ],
        "type": "object",
        "properties": {
          "identity": {
            "description": "The school identity group.",
            "$ref": "#/components/schemas/SettingsIdentityGroupDto"
          }
        },
        "additionalProperties": false,
        "description": "The response body of `GET /api/v1/settings`. Only SettingsIdentityGroupDto SettingsDto.Identity exists as of\nTASK-0005a; TASK-0005b and TASK-0005c extend this same envelope additively with sibling groups\n(logo/signature are read through SettingsIdentityGroupDto SettingsDto.Identity's own follow-up serving endpoints rather\nthan a new top-level field, and registration-number/abbreviation get their own group here).",
        "example": {
          "identity": {
            "schoolName": "Golden Royal Ark School",
            "shortName": "GRAS",
            "address": "12 Ark Crescent, Lekki, Lagos",
            "phone": "+2348012345678",
            "email": "info@goldenroyalark.example",
            "motto": "Excellence Through Character",
            "headTeacherName": "Chisom Maxwell",
            "timezone": "Africa/Lagos",
            "versionNumber": 3
          }
        }
      },
      "SettingsIdentityGroupDto": {
        "required": [
          "schoolName",
          "shortName",
          "address",
          "phone",
          "email",
          "motto",
          "headTeacherName",
          "timezone",
          "versionNumber"
        ],
        "type": "object",
        "properties": {
          "schoolName": {
            "type": "string",
            "description": "Full school name. Appears in full on the result sheet header.",
            "example": "Golden Royal Ark School"
          },
          "shortName": {
            "type": "string",
            "description": "Used where the full name will not fit, for example the pin slip.",
            "example": "GRAS"
          },
          "address": {
            "type": "string",
            "description": "Multi-line permitted.",
            "example": "12 Ark Crescent, Lekki, Lagos"
          },
          "phone": {
            "type": "string",
            "description": "Nigerian format, normalised to `+234` form.",
            "example": "+2348012345678"
          },
          "email": {
            "type": "string",
            "description": "Valid email format, stored lower-invariant.",
            "example": "info@goldenroyalark.example"
          },
          "motto": {
            "type": [
              "null",
              "string"
            ],
            "description": "`null` when unset. Printed under the school name if present.",
            "example": "Excellence Through Character"
          },
          "headTeacherName": {
            "type": "string",
            "description": "Printed above the head teacher's signature block.",
            "example": "Chisom Maxwell"
          },
          "timezone": {
            "type": "string",
            "description": "Always `Africa/Lagos`. Fixed; a `PATCH` cannot change it.",
            "example": "Africa/Lagos"
          },
          "versionNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "The identity group's current optimistic-concurrency pointer. Echo this back as\n`expectedVersion` on the next `PATCH`.",
            "format": "int32",
            "example": 3
          }
        },
        "additionalProperties": false,
        "description": "The school identity group, both inside SettingsDto and as\n`PATCH /api/v1/settings/identity`'s own success body (spec 6.2.3).",
        "example": {
          "schoolName": "Golden Royal Ark School",
          "shortName": "GRAS",
          "address": "12 Ark Crescent, Lekki, Lagos",
          "phone": "+2348012345678",
          "email": "info@goldenroyalark.example",
          "motto": "Excellence Through Character",
          "headTeacherName": "Chisom Maxwell",
          "timezone": "Africa/Lagos",
          "versionNumber": 3
        }
      },
...
      "UpdateSchoolIdentityCommand": {
        "required": [
          "schoolName",
          "shortName",
          "address",
          "phone",
          "email",
          "motto",
          "headTeacherName",
          "expectedVersion"
        ],
        "type": "object",
        "properties": {
          "schoolName": {
            "type": "string",
            "description": "Full school name.",
            "example": "Golden Royal Ark School"
          },
          "shortName": {
            "type": "string",
            "description": "Used where the full name will not fit.",
            "example": "GRAS"
          },
          "address": {
            "type": "string",
            "description": "Multi-line permitted.",
            "example": "12 Ark Crescent, Lekki, Lagos"
          },
          "phone": {
            "type": "string",
            "description": "Nigerian format — `08012345678` or `+2348012345678`.",
            "example": "08012345678"
          },
          "email": {
            "type": "string",
            "description": "Valid email format.",
            "example": "info@goldenroyalark.example"
          },
          "motto": {
            "type": [
              "null",
              "string"
            ],
            "description": "`null` to leave the school with no motto.",
            "example": "Excellence Through Character"
          },
          "headTeacherName": {
            "type": "string",
            "description": "Printed above the head teacher's signature block.",
            "example": "Chisom Maxwell"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "The identity group's current `versionNumber`, as last read from `GET /settings`. A\nstale value is rejected `409 settings.identity.stale_version` before anything is written.",
            "format": "int32",
            "example": 2
          }
        },
        "additionalProperties": false,
        "description": "`PATCH /api/v1/settings/identity` (spec 6.2.3). string SchoolProfile.Abbreviation and\n        string SchoolProfile.Timezone are deliberately absent from this body — the abbreviation is\n        TASK-0005c's own endpoint, and the timezone is fixed and not editable in this version.",
        "example": {
          "schoolName": "Golden Royal Ark School",
          "shortName": "GRAS",
          "address": "12 Ark Crescent, Lekki, Lagos",
          "phone": "08012345678",
          "email": "info@goldenroyalark.example",
          "motto": "Excellence Through Character",
          "headTeacherName": "Chisom Maxwell",
          "expectedVersion": 2
        }
      }
    },
    "securitySchemes": {
      "CookieSession": {
        "type": "apiKey",
        "description": "Opaque session token set by POST /api/v1/auth/sign-in. HttpOnly — the browser attaches it automatically; it is never readable or settable from JavaScript.",
        "name": "__Host-Session",
        "in": "cookie"
      }
    }
  },
...
}